// Home/transaction/fin-05a-invoice/model/fin-05a.ts
export class Fin05a {
    
    // stateName;
    // DistrictName;
    // clinicType;
    // clinic;
    stateId: number;
    districtId: number;
    clinicTypeId: number;
    // clinicCode: string;
    // clinicId: number;
    // month: string;
    // year: string;
    // assetName: string;
    typeOfExpenses: string;
    reference: string;

    siteConformityExpenses: number;
    // status: string;
    
}